# docs/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/users":{"get":{"operationId":"UsersController_getUsers","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]},"post":{"operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"put":{"operationId":"UsersController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/products":{"get":{"operationId":"ProductsController_getProducts","parameters":[],"responses":{"200":{"description":""}},"tags":["Products"]},"post":{"operationId":"ProductsController_createProduct","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":""}},"tags":["Products"]}},"/products/{id}":{"get":{"operationId":"ProductsController_getProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Products"]},"delete":{"operationId":"ProductsController_deleteProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Products"]},"put":{"operationId":"ProductsController_updateProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":""}},"tags":["Products"]}},"/auth/login":{"post":{"operationId":"AuthController_logIn","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/validate":{"post":{"operationId":"AuthController_validate","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}}},"info":{"title":"Products Backend","description":"Products Backend","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"CreateProductDto":{"type":"object","properties":{"Handle":{"type":"string"},"Title":{"type":"string"},"Description":{"type":"string"},"SKU":{"type":"string"},"Grams":{"type":"string"},"Stock":{"type":"string"},"Price":{"type":"string"},"ComparePrice":{"type":"string"},"Barcode":{"type":"string"}},"required":["Handle","Title","Description","SKU","Grams","Stock","Price","ComparePrice","Barcode"]},"UpdateProductDto":{"type":"object","properties":{"Handle":{"type":"string"},"Title":{"type":"string"},"Description":{"type":"string"},"SKU":{"type":"string"},"Grams":{"type":"string"},"Stock":{"type":"string"},"Price":{"type":"string"},"ComparePrice":{"type":"string"},"Barcode":{"type":"string"}}}}}}